Added a readme for task 5 and some changes
